requirements.txt
flask
flask-cors
requests
python-dotenv
pydantic
openai
google-genai
langgraph
